環境の修正
