+++ index.html
@@ -21,8 +21,9 @@
       src="https://unpkg.com/react-dom@18/umd/react-dom.development.js"
       crossorigin
     ></script>
+    <script src="https://unpkg.com/babel-standalone@6/babel.min.js"></script>
 
     <!-- Load our React component. -->
-    <script src="todo.js" type="module"></script>
+    <script src="src/todo.js" type="module"></script>
   </body>
 </html>
