--- package.json
@@ -1,22 +1,0 @@
-{
-  "name": "react-todo-app-practice",
-  "version": "1.0.0",
-  "description": "フィヨルドブートキャンプの「React の基本」プラクティス用リポジトリです。",
-  "main": "todo.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "rira100000000 <101010hayakawa@gmail.com> (https://github.com/rira100000000/npm_practice)",
-  "license": "ISC",
-  "dependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-preset-react-app": "^3.1.2",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  },
-  "devDependencies": {
-    "webpack": "^5.86.0",
-    "webpack-cli": "^5.1.4"
-  }
-}
